Add Elasticsearch as an option for searching support
This allows you to set an Elasticsearch endpoint to enable the search
API in Cadence.

It adds the following parameters to `values.yaml`:

| Parameter | Description | Default |
|-----------|-------------|--------|
| `elasticsearch.enabled` | Enable searching for workflows using an Elasticsearch cluster | `false` |
| `elasticsearch.scheme` | Scheme to use to connect to the Elasticsearch cluster | `http`|
| `elasticsearch.host` | Host to connect to | `` |
| `elasticsearch.indices` | Indices to use | `{}` |

These should match the relevant options that Cadence exposes for
configuring Elasticsearch.

Note that this does _not_ set up and configure an Elasticsearch
instance, but requires that you already have one set up and ready to go.

Closes #993.