feat(cadence): Preliminary PostgreSQL support
Updates the chart to better support a PostgreSQL backend.

This commit does not provide feature parity with existing Cassandra or
MySQL support.  The chart has options to spin up a MySQL or Cassandra
instance and configure it for you.  This commit offers nothing
equivalent for Postgres; if you want to use a Postgres backend you must
provision and configure it yourself.

More specifically, we assume that:
- The Postgres service is up and running.
- The `cadence` user and its password are configured.
- The databases `cadence` and `cadence_visibility` exist and are
  accessible by the `cadence` user.
(This assumes default names are used.  Adjust accordingly if you're
using different names.)

As long as those requirements are met, the jobs for "schema setup" and
"schema update" should work just as they would for the MySQL backend.
The option to disable either job through the "schema.setup.enabled"
and "schema.update.enabled" values remains as before.

The changes to support this new behavior are very minor.  We simply
remove an overly strict "if mysql" check in the setup job and add an
postgres implementation for the update job.