[prometheus-operator-standalone] Allow Helm to deploy CRDs too
Starting Helm v3 it is possible to deploy CRD manifests automatically when installing a chart, by simply placing manifests inside a folder named "crds".
As the chart provides the manifests already, all we have to do is rename the folder name in order to fit with the new Helm v3 workflow.

Fixes #1257